# Fix and improve apply_token_bitmask benchmark script